requirements.txt
# Requirements for Meal Nutrition Analyzer
streamlit>=1.22.0
pillow>=9.0.0
groq>=0.4.0
requests>=2.28.0
python-dotenv>=0.19.0
